Users,post and comment System all are working

+++ InstagramPost/Components/Post.cs
@@ -16,14 +16,11 @@ namespace InstagramPost.Components
 
         String Media { get; set; }
 
-        int Likes { get; set; }
-
-        bool IsLiked { get; set; }
-        bool IsDisLiked { get; set; }
-
+        HashSet<int> Likes { get; set; }
+        HashSet<int> Dislikes { get; set; } 
+        
         int TotalComments { get; set; }
-
-        int DisLikes { get; set; }
+        int CommentId { get; set; }
 
         List<Comment> Comments { get; set; }
 
@@ -32,29 +29,28 @@ namespace InstagramPost.Components
         public Post()
         {
             Comments = new List<Comment>();
-            Likes = 0;
-            DisLikes = 0;
-            IsLiked = false;
+            Likes=new HashSet<int>();
+            Dislikes=new HashSet<int>();
             TotalComments = 0;
+            CommentId = 0;
         }
 
         public Post(String media)
         {
             this.Media = media;
-            Likes = 0;
-            IsLiked=false;  IsDisLiked = false;
             Comments = new List<Comment>();
             CommentsWithID=new Dictionary<int,Comment>();
-            TotalComments = 0;
+            TotalComments = 0; CommentId = 0;
+            Likes=new HashSet<int>(); Dislikes = new HashSet<int>();
         }
 
         #region DisplayPost
-        public void DisplayPost()
+        public void DisplayPost(int PostIndex)
         {
-            Console.WriteLine("----------------------------------------------------------------------------------------------------\n\n");
+            Console.WriteLine("-------------------------------------- PostIndex : "+PostIndex+" -----------------------------------------------\n\n");
             Console.WriteLine("\t\t\t\tPost : " + this.Media);
-            Console.WriteLine("\t\t\t\tTotal Likes : " + this.Likes);
-            Console.WriteLine("\t\t\t\tTotal DisLikes : " + this.DisLikes);
+            Console.WriteLine("\t\t\t\tTotal Likes : " + Likes.Count);
+            Console.WriteLine("\t\t\t\tTotal DisLikes : " + Dislikes.Count);
             Console.WriteLine("\n\nComments : " + TotalComments + "\n");
 
             foreach (Comment comment in Comments)
@@ -64,7 +60,7 @@ namespace InstagramPost.Components
 
             Console.WriteLine("----------------------------------------------------------------------------------------------------");
 
-            Reaction();
+           // Reaction(PostIndex);
 
         }
         #endregion
@@ -75,7 +71,7 @@ namespace InstagramPost.Components
             Console.WriteLine("0. Exit               1. Add Comment          2. Add Reply");
             Console.WriteLine("3. Like Post          4.DisLike Post          5. DeleteComment");
             Console.WriteLine("6. Like Comment       7.DisLike Comment       8. EditComment");
-            Console.Write("Select Your PostReaction Decision [0-5] : ");
+            Console.Write("Select Your PostReaction Decision [0-8] : ");
 
             int decisionIndex = -1;
             try
@@ -100,60 +96,62 @@ namespace InstagramPost.Components
         #endregion
 
         #region Reaction
-        public void Reaction()
+        public void Reaction(int PostIndex,int UserId)
         {
             int decisionIndex = Decision();
 
             switch (decisionIndex)
             {
                 case 1:
-                        AddComment(GenerateComment(0));   break;
+                        AddComment(GenerateComment(0,UserId));   break;
                 case 2:
-                        AddReply(); break;
+                        AddReply(UserId); break;
                 case 3:
-                        Like();     break;
+                        Like(UserId);     break;
                 case 4:
-                        DisLike();  break;
+                        DisLike(UserId);  break;
                 case 5:
-                        DeleteComment();   break;
+                        DeleteComment(UserId);   break;
                 case 6:
-                    LikeComment(); break;
+                        LikeComment(UserId); break;
                 case 7:
-                    DisLikeComment(); break;
+                        DisLikeComment(UserId); break;
                 case 8:
-                    EditComment(); break;
+                        EditComment(UserId); break;
             }
 
             if (decisionIndex != 0)
             {
                 Console.Clear();
-                DisplayPost();
+               // DisplayPost(PostIndex);
             }
 
         }
         #endregion
 
         #region Like Comment
-        public void LikeComment()
+        public void LikeComment(int UserId)
         { 
             Comment comment = GetComment();
-            comment.Like();
+            comment.Like(UserId);
         }
         #endregion
 
         #region DisLike Comment
-        public void DisLikeComment()
+        public void DisLikeComment(int UserId)
         {
             Comment comment = GetComment();
-            comment.DisLike();
+            comment.DisLike(UserId);
         }
         #endregion
 
         #region Edit Comment
-        public void EditComment()
+        public void EditComment(int UserId)
         {
             Comment comment = GetComment();
-            comment.EditComment();
+
+            if(comment.GetCommentID()!=-1)
+                comment.EditComment(UserId);
         }
         #endregion
 
@@ -184,18 +182,18 @@ namespace InstagramPost.Components
             {
                 return CommentsWithID[CommentId];
             }
-            else
+            else if(CommentId!=-1)
             {
                 Console.WriteLine("this Comment is not Exist");     //error msg
                 GetComment();
             }
 
-            return new Comment();
+            return new Comment(-1);
         }
         #endregion
 
         #region Add Reply    
-        public void AddReply()
+        public void AddReply(int UserId)
         {
                 Comment comment = GetComment();
 
@@ -203,12 +201,12 @@ namespace InstagramPost.Components
 
                 if(Hierarchy > 4) 
                 {
-                    Console.WriteLine("you can't add hirarchycal 6th comment...");  //error msg
+                    Console.WriteLine("you can't add hierarchical 6th comment...");  //error msg
                     Console.ReadLine();
                 }
                 else if (!comment.GetIsDeleted()) 
                 {
-                    comment.ReplyofComment(GenerateComment(Hierarchy));
+                    comment.ReplyofComment(GenerateComment(Hierarchy,UserId));
                     comment.SetHasReply(true);
                 }
                 else
@@ -220,51 +218,60 @@ namespace InstagramPost.Components
         #endregion
 
         #region DeleteComment
-        public void DeleteComment()
+        public void DeleteComment(int UserId)
         {
             int CommentId = 0;
-
-            Console.Write("Enter CommentId : ");
-            try
+            bool IsValidCommentId = true;
+            do
             {
-                CommentId = Int32.Parse(Console.ReadLine());
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine(ex.Message);
-                DeleteComment();
-            }
+                Console.Write("Enter CommentId : ");
+                try
+                {
+                    CommentId = Int32.Parse(Console.ReadLine());
+                    IsValidCommentId = true;
+                }
+                catch (Exception ex)
+                {
+                    Console.WriteLine(ex.Message);
+                    IsValidCommentId= false;
+                }
+            } while (!IsValidCommentId);
                 
                 if (!(CommentsWithID.ContainsKey(CommentId))) { Console.WriteLine("Comment Not Exist."); Console.ReadLine(); }
                 else
                 {
-                
                     Comment comment = CommentsWithID[CommentId];
-
-                    if(comment.GetIsDeleted())
-                    { 
-                        Console.WriteLine("Comment allready Deleted");
-                        Console.ReadLine();
-                    }
-                    else if(comment.GetHasReply()) 
+                    if (comment.GetUserId() == UserId)
                     {
-                        comment.DeleteValues();
-                        comment.SetIsDeleted(true);
-                        TotalComments--;
+                        if (comment.GetIsDeleted())
+                        {
+                            Console.WriteLine("Comment allready Deleted");
+                            Console.ReadLine();
+                        }
+                        else if (comment.GetHasReply())
+                        {
+                            comment.DeleteValues();
+                            comment.SetIsDeleted(true);
+                            TotalComments--;
+                        }
+                        else
+                        {
+                            Comments.Remove(comment);
+                            TotalComments--;
+                        }
                     }
                     else
                     {
-                        Comments.Remove(comment);
-                        TotalComments--;
+                        Console.WriteLine("This is Not Your Comment so you can't Delete it");
+                        Console.ReadLine();
                     }
-
                 }
 
         }
 
         #endregion
 
-        #region Get value Methods
+        #region GetvalueMethods
         public int GetTotalComments()
         {
             return this.TotalComments;
@@ -272,7 +279,7 @@ namespace InstagramPost.Components
         #endregion
 
         #region GenerateComment
-        public Comment GenerateComment(int Hierarchy)
+        public Comment GenerateComment(int Hierarchy,int UserId)
         {
             String CommentText = "Initial Comment";
             Console.Write("Your Comment : ");
@@ -281,13 +288,14 @@ namespace InstagramPost.Components
             {
                 CommentText = Console.ReadLine();
                 TotalComments++;
+                
             }
             catch (Exception ex)
             {
                 Console.WriteLine(ex.Message);
             }
 
-            Comment comment= new Comment(CommentText, this, Hierarchy);
+            Comment comment= new Comment(CommentText, CommentId++, Hierarchy,UserId);
             CommentsWithID.Add(comment.GetCommentID(), comment);
 
             return comment;
@@ -303,43 +311,22 @@ namespace InstagramPost.Components
         #endregion
 
         #region Like Post
-        public void Like()  
+        public void Like(int UserId)  
         {
-            //Likes++;
+            if(!Likes.Contains(UserId))
+                Likes.Add(UserId);
             
-            if (!IsLiked)
-            {
-                Likes++;
-                IsLiked = true;
-
-            } 
-
-            if(IsDisLiked)
-            {
-                DisLikes--;
-                IsDisLiked = false;
-            }
-            
-        
+            Dislikes.Remove(UserId);
         }
         #endregion
 
         #region DisLike Post
-        public void DisLike() 
+        public void DisLike(int UserId) 
         {
-            //DisLikes++;
-            
-            if (!IsDisLiked)
-            {
-                DisLikes++;
-                IsDisLiked = true;
-            }
-            if(IsLiked)
-            {
-                Likes--;
-                IsLiked = false;
+            if (!Dislikes.Contains(UserId))
+                Dislikes.Add(UserId);
 
-            }  
+            Likes.Remove(UserId);
         }
         #endregion
